64-bit support

+++ Provenance/zipkit/NSData+ZKAdditions.h
@@ -9,15 +9,15 @@
 
 @interface NSData (ZKAdditions)
 
-- (UInt16) zk_hostInt16OffsetBy:(NSUInteger *)offset;
-- (UInt32) zk_hostInt32OffsetBy:(NSUInteger *)offset;
-- (UInt64) zk_hostInt64OffsetBy:(NSUInteger *)offset;
-- (BOOL) zk_hostBoolOffsetBy:(NSUInteger *) offset;
-- (NSString *) zk_stringOffsetBy:(NSUInteger *)offset length:(NSUInteger)length;
-- (NSUInteger) zk_crc32;
-- (NSUInteger) zk_crc32:(NSUInteger)crc;
-- (NSData *) zk_inflate;
-- (NSData *) zk_deflate;
+- (UInt16) zk_hostInt16OffsetBy:(UInt64 *)offset;
+- (UInt32) zk_hostInt32OffsetBy:(UInt64 *)offset;
+- (UInt64) zk_hostInt64OffsetBy:(UInt64 *)offset;
+- (BOOL) zk_hostBoolOffsetBy:(UInt64 *)offset;
+- (NSString *) zk_stringOffsetBy:(UInt64 *)offset length:(NSUInteger)length;
+- (unsigned long)       zk_crc32;
+- (unsigned long) zk_crc32:(unsigned long)crc;
+- (NSData *)            zk_inflate;
+- (NSData *)            zk_deflate;
 
 @end
 
@@ -30,7 +30,7 @@
 - (void) zk_appendLittleInt16:(UInt16)value;
 - (void) zk_appendLittleInt32:(UInt32)value;
 - (void) zk_appendLittleInt64:(UInt64)value;
-- (void) zk_appendLittleBool:(BOOL) value;
+- (void) zk_appendLittleBool:(BOOL)value;
 - (void) zk_appendPrecomposedUTF8String:(NSString *)value;
 
 @end
